requirements.txt
pyyaml
yapf
python-configuration[toml,yaml]
requests
beautifulsoup4
pytest